Module askama_shared::filters

Module for built-in filter functions

Contains all the built-in filter functions for use in templates. You can define your own filters, as well. For more information, read the book.

Constants

BUILT_IN_FILTERS

Functions

capitalize

Capitalize a value. The first character will be uppercase, all others lowercase.

center

Centers the value in a field of a given width

escape

Escapes &, < and > in strings

fmt

Formats arguments according to the specified format

format

Formats arguments according to the specified format

indent

Indent lines with width spaces

join

Joins iterable into a string separated by provided argument

linebreaks

Replaces line breaks in plain text with appropriate HTML

linebreaksbr

Converts all newlines in a piece of plain text to HTML line breaks

lower

Converts to lowercase

lowercase

Alias for the lower() filter

paragraphbreaks

Replaces only paragraph breaks in plain text with appropriate HTML

safe

Marks a string (or other Display type) as safe

trim

Strip leading and trailing whitespace

truncate

Limit string length, appends ‘…’ if truncated

upper

Converts to uppercase

uppercase

Alias for the upper() filter

wordcount

Count the words in that string